pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-detector"
version = "0.0.1"
description = "A general-purpose engine that detects and masks personal information by country and information type"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name="zafrem", email="zafrem@gmail.com"}
]
keywords = ["pii", "detection", "redaction", "regex", "privacy", "gdpr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "click>=8.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0",
    "prometheus-client>=0.19.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pyyaml",
    "types-jsonschema",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
data-detector = "datadetector.cli:main"

[project.urls]
Homepage = "https://github.com/zafrem/data-detector"
Repository = "https://github.com/zafrem/data-detector"
Issues = "https://github.com/zafrem/data-detector/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
datadetector = ["py.typed", "schemas/*.json"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
]
# Ignore Python 3.9+ specific rules (project supports Python 3.8+)
ignore = [
    "UP006",  # Use builtin collection types (list, dict, tuple) - requires Python 3.9+
    "UP035",  # Import from typing instead of collections.abc - requires Python 3.9+
]

[tool.mypy]
python_version = "3.8"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/datadetector"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
